+++ Packages/com.texelsaur.video/Runtime/Scripts/Component/URL Source/PlaylistQueue.cs
@@ -20,6 +20,11 @@ namespace Texel
     {
         TXLVideoPlayer videoPlayer;
 
+        [SerializeField] protected internal bool limitEntries = true;
+        [SerializeField] protected internal int maxEntries = 30;
+        [SerializeField] protected internal bool limitEntriesPerPlayer = false;
+        [SerializeField] protected internal int maxEntriesPerPlayer = 0;
+
         [SerializeField] protected internal bool allowAdd = true;
         [SerializeField] protected internal AccessControl addAccess;
         [SerializeField] protected internal bool allowAddFromProxy = false;
@@ -29,15 +34,18 @@ namespace Texel
         [SerializeField] protected internal bool allowPriority = false;
         [Tooltip("Optional. ACL to control access to the delete button.  If not set, uses the video player's ACL settings.")]
         [SerializeField] protected internal AccessControl deleteAccess;
-        //[Tooltip("Allows players to delete their own added entries, even if th")]
+        //[Tooltip("Allows players to delete their own added entries, even if th")] 
         [SerializeField] protected internal bool allowDelete = true;
         [SerializeField] protected internal bool allowSelfDelete = false;
+        [SerializeField] protected internal AccessControl moveAccess;
+        [SerializeField] protected internal bool allowMove = false;
 
         [SerializeField] protected internal bool removeOnLeave = false;
         [SerializeField] protected internal bool removeIfAbsent = false;
 
         [SerializeField] protected internal bool canInterruptSources = true;
 
+        [SerializeField] protected internal bool transferOwnershipOnPlay = false;
         [SerializeField] protected internal bool enableSyncQuestUrls = true;
         [SerializeField] protected internal bool syncTrackTitles = true;
         [SerializeField] protected internal bool syncTrackAuthors = true;
@@ -56,6 +64,8 @@ namespace Texel
         VRCUrl syncReadyQuestUrl;
         [UdonSynced]
         string syncReadyTitle;
+        [UdonSynced]
+        int syncReadyPlayerId;
         [UdonSynced]
         string syncReadyPlayer;
         [UdonSynced]
@@ -127,6 +137,7 @@ namespace Texel
 
             syncReadyUrl = VRCUrl.Empty;
             syncReadyTitle = "";
+            syncReadyPlayerId = -1;
             syncReadyPlayer = "";
 
             syncUrls = new VRCUrl[0];
@@ -316,6 +327,20 @@ namespace Texel
             }
         }
 
+        public bool HasMoveAccess
+        {
+            get
+            {
+                if (!allowMove)
+                    return false;
+                if (moveAccess)
+                    return moveAccess._LocalHasAccess();
+                if (videoPlayer)
+                    return videoPlayer._CanTakeControl();
+                return false;
+            }
+        }
+
         public bool _HasDeleteAccessFor(int index)
         {
             if (!allowDelete || index < 0 || index >= syncTrackCount)
@@ -371,7 +396,12 @@ namespace Texel
             return syncReadyTitle;
         }
 
-        public virtual string _GetCurrentPlayer()
+        public override int _GetCurrentPlayerId()
+        {
+            return syncReadyPlayerId;
+        }
+
+        public override string _GetCurrentPlayerName()
         {
             return syncReadyPlayer;
         }
@@ -448,6 +478,11 @@ namespace Texel
             return "";
         }
 
+        public int _GetTrackPlayerId(int index)
+        {
+            return syncPlayerIds[index];
+        }
+
         public string _GetTrackPlayer(int index)
         {
             if (!usingPlayers)
@@ -486,7 +521,7 @@ namespace Texel
 
         public override bool _MoveNext()
         {
-            if (!_TakeControl())
+            if (!_ForceTakeControl())
                 return false;
 
             if (syncTrackCount == 0)
@@ -519,18 +554,30 @@ namespace Texel
             syncReadyUrl = _GetTrackURL(0);
             syncReadyQuestUrl = _GetTrackURL(0, TXLUrlType.Quest);
             syncReadyTitle = _GetTrackName(0);
+            syncReadyPlayerId = _GetTrackPlayerId(0);
             syncReadyPlayer = _GetTrackPlayer(0);
+            int readyPlayerId = syncPlayerIds[0];
 
             _PopTracks(0, 1);
-
-            _EventTrackChange();
-            _EventUrlReady();
 
             errorCount = 0;
             _IncrReadyUrlUpdate();
             _IncrTrackChangeUpdate();
 
             RequestSerialization();
+
+            if (transferOwnershipOnPlay)
+            {
+                VRCPlayerApi player = VRCPlayerApi.GetPlayerById(readyPlayerId);
+                if (Utilities.IsValid(player))
+                {
+                    if (usingDebug) _DebugLog($"Transfering video player ownership to {player.displayName} [{readyPlayerId}]");
+                    Networking.SetOwner(player, videoPlayer.gameObject);
+                }
+            }
+
+            _EventTrackChange();
+            _EventUrlReady();
 
             return true;
         }
@@ -684,19 +731,27 @@ namespace Texel
         {
             if (!allowPriority)
                 return false;
-            if (!_TakeControl(priorityAccess))
-                return false;
 
-            return _MoveTrack(index, 0);
+            return _MoveTrackLL(index, 0, priorityAccess);
         }
 
         public bool _MoveTrack(int index, int destIndex)
+        {
+            if (!allowMove)
+                return false;
+
+            return _MoveTrackLL(index, destIndex, moveAccess);
+        }
+
+        bool _MoveTrackLL(int index, int destIndex, AccessControl access)
         {
             if (index < 0 || index >= syncTrackCount)
                 return false;
             if (destIndex < 0 || destIndex >= syncTrackCount)
                 return false;
             if (index == destIndex)
+                return false;
+            if (!_TakeControl(access))
                 return false;
 
             if (usingDebug) _DebugLog($"Move track from {index} to {destIndex}");
@@ -773,6 +828,25 @@ namespace Texel
 
         public override bool _CanAddTrack()
         {
+            if (!allowAdd)
+                return false;
+            if (limitEntries && syncTrackCount >= maxEntries)
+                return false;
+
+            if (limitEntriesPerPlayer)
+            {
+                int playerId = Networking.LocalPlayer.playerId;
+                int foundCount = 0;
+                for (int i = 0; i < syncTrackCount; i++)
+                {
+                    if (syncPlayerIds[i] == playerId)
+                        foundCount += 1;
+                }
+
+                if (foundCount >= maxEntriesPerPlayer)
+                    return false;
+            }
+
             return true;
         }
 
@@ -793,7 +867,7 @@ namespace Texel
 
         public bool _AddTrack(VRCUrl url, VRCUrl questUrl, string title, string author, bool force)
         {
-            if (!allowAdd)
+            if (!_CanAddTrack())
                 return false;
             if (!URLUtil.WellFormedUrl(url))
                 return false;
@@ -916,6 +990,8 @@ namespace Texel
 
         public bool _AddTrack(int playlistIndex, int catalogIndex, int trackIndex)
         {
+            if (!_CanAddTrack())
+                return false;
             if (!_TakeControl(addAccess))
                 return false;